requirements.txt
cryptography
fire
install_playwright
jsonschema
numpy<=2.2.6
pandas
pillow
playwright>=1.54.0,<=1.56.0
pydantic>=2.10.6
requests
retry
tenacity
pytz
strands-agents>=1.2.0,<=1.14.0
strands-agents-tools<=0.2.13
Deprecated
boto3>=1.42.1
boto3-stubs
mypy-boto3-s3
